refactor: modify float to double for order entity & order item entity

+++ order-service/src/main/java/com/nadiannis/order_service/service/OrderService.java
@@ -84,7 +84,7 @@ public class OrderService {
         if (orderReqDto.getOrderItems() == null) {
             orderReqDto.setOrderItems(Collections.emptyList());
         }
-        Float totalAmount = (float) orderReqDto.getOrderItems().stream()
+        Double totalAmount = orderReqDto.getOrderItems().stream()
                 .mapToDouble(orderItemReqDto -> orderItemReqDto.getPrice() * orderItemReqDto.getQuantity())
                 .sum();
